requirements.txt
freetype-py==2.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycairo==1.23.0
reportlab==4.0.0
rlPyCairo==0.2.0
